Cranked up warnings even higher. MS headers fall over, trying experimental switch to give them a different level. Have signed/unsigned mismatch disabled with 4365 and 4287, need to fix that later.

+++ libs/core/rf_assert.h
@@ -24,7 +24,7 @@ AssertResponse AssertNotification( char const* file, size_t line, char const* fa
 		if( \
 			___rf_assert_skip_all_instances == false && \
 			!!!(TEST) && \
-			::RF::assert::AssertNotification(__FILE__, __LINE__, #TEST, MESSAGEEXPRESSION) \
+			::RF::assert::AssertNotification(__FILE__, static_cast<size_t>(__LINE__), #TEST, MESSAGEEXPRESSION) \
 				== ::RF::assert::AssertResponse::Interrupt ) \
 		{ \
 			RF_SOFTWARE_INTERRUPT(); \
